% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_parser.R, R/zzz-aliases.R
\name{classify_last_names}
\alias{classify_last_names}
\alias{ent_classify_ethnicity}
\title{Classify last names}
\usage{
classify_last_names(
  last_names = c("Jackson", "Cheng", "Wen"),
  include_probabilities = F,
  last_name_column = "nameLast",
  snake_names = F,
  return_message = T
)

ent_classify_ethnicity(
  last_names = c("Jackson", "Cheng", "Wen"),
  include_probabilities = F,
  last_name_column = "nameLast",
  snake_names = F,
  return_message = T
)
}
\arguments{
\item{last_names}{vector of last names}

\item{include_probabilities}{if \code{TRUE} appends race probabilities}

\item{last_name_column}{column name for last names in output}

\item{snake_names}{if \code{TRUE} converts column names to snake_case}

\item{return_message}{if \code{TRUE} prints progress messages}
}
\value{
A tibble with surname and predicted race probabilities
}
\description{
Classify a vector of last names using the WRU (Who Are You) model
to predict race/ethnicity probabilities.
}
\examples{
\dontrun{
classify_last_names(c("Jackson", "Cheng", "Wen"))
}
}
