% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_resolvers.R
\name{dictionary_entity_slugs}
\alias{dictionary_entity_slugs}
\title{Entity dictionary words}
\usage{
dictionary_entity_slugs()
}
\value{
a \code{tibble}
}
\description{
Dictionary of words that indicate entity status
}
\examples{
dictionary_entity_slugs()
}
