% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{dictionary_fb_countries}
\alias{dictionary_fb_countries}
\title{Forebears country dictionary}
\usage{
dictionary_fb_countries()
}
\value{
A tibble with columns: idCountry, nameCountry, codeCountry, isDisputed
}
\description{
Returns a tibble of countries and their metadata from Forebears.io GeoJSON data.
}
\examples{
\dontrun{
dictionary_fb_countries()
}
}
