% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{.validate_rules}
\alias{.validate_rules}
\title{Validate entity_rules object}
\usage{
.validate_rules(rules)
}
\arguments{
\item{rules}{An entity_rules object}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Validate entity_rules object
}
\keyword{internal}
