% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities-package.R
\docType{package}
\name{entities-package}
\alias{entities}
\alias{entities-package}
\title{entities: Tools to deal with entities}
\description{
Clean, resolve and parse your entities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abresler/entities}
}

}
\author{
\strong{Maintainer}: Alex Bresler \email{abresler@asbcllc.com}

}
\keyword{internal}
