% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{entity_abbreviations}
\alias{entity_abbreviations}
\title{Entity Abbreviations}
\usage{
entity_abbreviations(case = "upper", remove_commas = T, remove_periods = T)
}
\arguments{
\item{remove_periods}{}
}
\description{
Entity Abbreviations
}
