% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{fb_last_names}
\alias{fb_last_names}
\title{Forebears last name data}
\usage{
fb_last_names(
  last_names = NULL,
  sleep_time = NULL,
  include_similar_names = F,
  return_message = T,
  nest_data = F,
  snake_names = F,
  assign_to_environment = T
)
}
\arguments{
\item{last_names}{vector of last names}

\item{sleep_time}{if not \code{NULL} time between queries}

\item{include_similar_names}{if \code{TRUE} includes table about similar names}

\item{return_message}{if \code{TRUE} returns message}

\item{nest_data}{if \code{TRUE} nests tibble}

\item{snake_names}{if \code{TRUE} converts column names to snake_case}

\item{assign_to_environment}{if \code{TRUE} assigns nested tibbles to environment}
}
\value{
A tibble with surname data, or empty tibble if data unavailable
}
\description{
Retrieves surname demographic data from forebears.io. Note that this function
may fail if the website has changed its structure or blocked automated access.
}
\examples{
\dontrun{
fb_last_names(last_names = "Smith")
}
}
