% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{fb_locations}
\alias{fb_locations}
\title{Forebears country data}
\usage{
fb_locations(locations = NULL, snake_names = F, return_message = T)
}
\arguments{
\item{locations}{vector of countries}

\item{return_message}{if \code{TRUE} returns message}
}
\value{

}
\description{
Forebears country data
}
\examples{
fb_locations(locations = c("China"))
}
