% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{print.entity_resolution}
\alias{print.entity_resolution}
\title{Print method for entity_resolution objects}
\usage{
\method{print}{entity_resolution}(x, ...)
}
\arguments{
\item{x}{An entity_resolution object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for entity_resolution objects
}
