% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{print.entity_rules}
\alias{print.entity_rules}
\title{Print method for entity_rules objects}
\usage{
\method{print}{entity_rules}(x, ...)
}
\arguments{
\item{x}{An entity_rules object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for entity_rules objects
}
