% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_reresolver.R
\name{re_resolve_documents}
\alias{re_resolve_documents}
\title{Re - Resolve ACRIS Parties}
\usage{
re_resolve_documents(
  base_path = "Desktop/nyc_real_estate_data/ACRIS/",
  documents = "ucc3 assignment",
  boroughs = c("brooklyn", "manhattan")
)
}
\arguments{
\item{base_path}{base ACRIS data path}

\item{documents}{vector of document types to process}

\item{boroughs}{vector of NYC boroughs to process}
}
\value{
invisible (side effects: writes cleaned data to files)
}
\description{
Re-resolves and cleans party data from ACRIS (Automated City Register
Information System) documents.
}
\examples{
\dontrun{
re_resolve_documents(documents = "ucc3 assignment", boroughs = "manhattan")
}
}
