% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinr.R
\name{refine_columns}
\alias{refine_columns}
\title{Refine entity columns}
\usage{
refine_columns(data, entity_columns = NULL, use_business_suffix = T,
  phonics_methods = "soundex", phonics_length = 16L,
  use_n_gram_merge = T, edit_threshold = 1, ignore_words = NULL,
  numgram = 2, weight = c(d = 0.33, i = 0.33, s = 1, t = 0.5), ...)
}
\arguments{
\item{...}{}
}
\description{
Refine entity columns
}
