% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_resolvers.R
\name{resolve_data_parties}
\alias{resolve_data_parties}
\title{Resolve List of Parties}
\usage{
resolve_data_parties(
  parties = c("LIU, YUAN-CHI", "ASBC LLC"),
  entity_words = NULL,
  remove_first_last = TRUE,
  variable = "nameOwnerPrimary",
  snake_names = F
)
}
\arguments{
\item{parties}{vector of parties}

\item{entity_words}{dictionary of entity words coming from \code{dictionary_entity_slugs}}

\item{remove_first_last}{if \code{TRUE} removes first and last name}

\item{variable}{output variable}

\item{snake_names}{if \code{TRUE} converts column names to snake_case (default FALSE)}
}
\value{
a \code{tibble}
}
\description{
Resolve List of Parties
}
\examples{
resolve_data_parties(parties = c("LIU, YUAN-CHI", "ASBC LLC"),
remove_first_last = TRUE,
variable = "nameOwnerPrimary")
}
