% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_distance_join}
\alias{tbl_distance_join}
\title{Join two tables based on a distance metric of one or more columns
This differs from difference_join in that it considers all of the columns together when computing distance. This allows it to use metrics such as Euclidean or Manhattan that depend on multiple columns. Note that if you are computing with longitude or latitude, you probably want to use geo_join.}
\usage{
tbl_distance_join(
  x,
  y,
  by = NULL,
  max_dist = 1,
  methods = c("euclidean", "manhattan"),
  mode = "inner",
  distance_col = NULL
)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns by which to join the two tables}

\item{max_dist}{Maximum distance to use for joining}

\item{methods}{Method to use for computing distance, either `euclidean` (default) or `manhattan.`}

\item{mode}{One of "inner", "left", "right", "full" "semi", or "anti"}

\item{distance_col}{If given, will add a column with this name containing the distance between the two}
}
\value{

}
\description{
Join two tables based on a distance metric of one or more columns
This differs from difference_join in that it considers all of the columns together when computing distance. This allows it to use metrics such as Euclidean or Manhattan that depend on multiple columns. Note that if you are computing with longitude or latitude, you probably want to use geo_join.
}
