% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_names.R
\name{tbl_extract_human_name_parts}
\alias{tbl_extract_human_name_parts}
\title{Extract human name parts from a data frame}
\usage{
tbl_extract_human_name_parts(data, name_columns)
}
\arguments{
\item{data}{a data frame or tibble containing human names}

\item{name_columns}{a character vector of column names containing human names to be processed}
}
\description{
Extract human name parts from a data frame
}
