% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_parser.R
\name{tbl_last_name}
\alias{tbl_last_name}
\title{Parse name column into parts}
\usage{
tbl_last_name(data, name_column = "namePrincipalInvestigator",
  include_name_type = F, return_only_names = F)
}
\arguments{
\item{return_only_names}{}
}
\description{
Parse name column into parts
}
\examples{
library(dplyr)
starwars \%>\% tbl_last_name(name_column = "name")
}
