% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_names.R
\name{tbl_munge_human_names}
\alias{tbl_munge_human_names}
\title{Name Parser}
\usage{
tbl_munge_human_names(data, first_name, last_name, name_column)
}
\arguments{
\item{data}{a data frame or tibble containing human names}

\item{first_name}{vector referencing first names}

\item{last_name}{vector referencing last names}

\item{name_column}{name of the `name` you wish to create}
}
\description{
Name Parser
}
