% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{us_religions}
\alias{us_religions}
\title{Forebears US Religion data}
\usage{
us_religions(snake_names = F)
}
\arguments{
\item{snake_names}{if \code{TRUE} converts column names to snake_case}
}
\value{
A tibble with religion data by surname, or NULL if the data source
  is unavailable
}
\description{
Forebears US Religion data
}
\examples{
\dontrun{
us_religions()
}
}
