% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_parser.R
\name{classify_last_names}
\alias{classify_last_names}
\title{Classify last names}
\usage{
classify_last_names(last_names = c("Jackson", "Cheng", "Wen"),
  include_probabilities = F, last_name_column = "nameLast",
  return_message = T)
}
\arguments{
\item{last_names}{vector of last names}

\item{include_probabilities}{if \code{TRUE} appends probabilities}

\item{last_name_column}{if not \code{NULL} last name column}

\item{return_message}{}
}
\description{
Classify a vectore last names using the WRU model
}
\examples{
classify_last_names(last_names = c("Jackson", "Cheng", "Wen"), include_probabilities = F)
}
