% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_parser.R
\name{classify_wru_names}
\alias{classify_wru_names}
\title{Classify name columns for WRU ethnicity}
\usage{
classify_wru_names(data, name_columns = NULL,
  include_probabilities = F, include_name_type = T,
  return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Classify name columns for WRU ethnicity
}
