% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionary_countries_legal_entity_types}
\alias{dictionary_countries_legal_entity_types}
\title{Returns descriptions for countries corporate entity types}
\usage{
dictionary_countries_legal_entity_types(case = "upper", remove_periods = F)
}
\arguments{
\item{case}{text case \itemize{
\item `NULL`
\item `lower`
\item `lower`
}}

\item{remove_periods}{}
}
\value{
\code{tibble}
}
\description{
Returns descriptions for countries corporate entity types
}
