% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{dictionary_fb_countries}
\alias{dictionary_fb_countries}
\title{Forebears country dictionary}
\usage{
dictionary_fb_countries()
}
\description{
Forebears country dictionary
}
\examples{
dictionary_fb_countries()
}
