% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionary_gleif_entity_types}
\alias{dictionary_gleif_entity_types}
\title{GLEIF entity list}
\usage{
dictionary_gleif_entity_types(
  url =
    "https://www.gleif.org/about-lei/code-lists/iso-20275-entity-legal-forms-code-list/2023-09-28-elf-code-list-v1.5.csv"
)
}
\arguments{
\item{url}{url with most recent CSV}
}
\description{
Dictionary of GLEIF entities
}
