% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_engine.R
\name{query_words}
\alias{query_words}
\title{Query Search engies}
\usage{
query_words(
  engines = c("ddg", "google", "baidu", "startpage", "yandex"),
  words = NULL,
  quote = T,
  rename_output = NULL,
  additional_terms = NULL,
  return_message = T
)
}
\arguments{
\item{engines}{vector of search engines to search \itemize{
\item google
\item yandex
\item startpage
\item ddg
\item baidu
}}

\item{return_message}{}
}
\value{

}
\description{
Query Search engies
}
