% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_reresolver.R
\name{re_resolve_documents}
\alias{re_resolve_documents}
\title{Re - Resolve ACRIS Parties}
\usage{
re_resolve_documents(base_path = "Desktop/nyc_real_estate_data/ACRIS/",
  documents = "ucc3 assignment", boroughs = c("brooklyn", "manhattan"))
}
\arguments{
\item{base_path}{base ACRIS data path}

\item{documents}{vector of document types}

\item{boroughs}{vector of boroughs}
}
\value{
invisible
}
\description{
Re - Resolve ACRIS Parties
}
