% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_clean_variables}
\alias{tbl_clean_variables}
\title{Clean Variables}
\usage{
tbl_clean_variables(
  data,
  variables = NULL,
  all_character_columns = F,
  exclude_url_columns = T,
  case = "upper",
  split_characters = NULL,
  remove_bracketed_text = T,
  remove_digits = F,
  remove_commas = T,
  remove_periods = T,
  remove_html = T,
  remove_dash = F,
  remove_kern = F,
  remove_symbols = F,
  fix_comma_space = T,
  remove_ampersand = F,
  remove_entity_types = F,
  snake_names = T,
  overwrite_variables = F,
  ...
)
}
\arguments{
\item{overwrite_variables}{}
}
\value{

}
\description{
Clean Variables
}
