% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_difference_join}
\alias{tbl_difference_join}
\title{Join two tables based on absolute difference between their columns}
\usage{
tbl_difference_join(
  x,
  y,
  by = NULL,
  max_dist = 1,
  mode = "inner",
  distance_col = NULL
)
}
\arguments{
\item{distance_col}{}
}
\value{

}
\description{
Join two tables based on absolute difference between their columns
}
