% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_fuzzy_join}
\alias{tbl_fuzzy_join}
\title{Fuzzy Join}
\usage{
tbl_fuzzy_join(
  x,
  y,
  by = NULL,
  match_fun = NULL,
  multi_by = NULL,
  multi_match_fun = NULL,
  index_match_fun = NULL,
  mode = "inner",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Fuzzy Join
}
