% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics.R
\name{tbl_phonics}
\alias{tbl_phonics}
\title{Append phonics methods to a tbl}
\usage{
tbl_phonics(
  data,
  name_columns = NULL,
  phonics_methods = c("soundex"),
  phonics_length = 12L
)
}
\arguments{
\item{phonics_length}{}
}
\value{

}
\description{
Append phonics methods to a tbl
}
