% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_regex_join}
\alias{tbl_regex_join}
\title{Regex Join}
\usage{
tbl_regex_join(x, y, by = NULL, mode = "inner", ignore_case = FALSE)
}
\arguments{
\item{ignore_case}{}
}
\value{

}
\description{
Regex Join
}
