% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_variable_stringdist}
\alias{tbl_variable_stringdist}
\title{Tidy String Distance Calculation from Variable}
\usage{
tbl_variable_stringdist(
  data,
  variables = NULL,
  method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  clean_names = T,
  return_wide = T,
  ...
)
}
\arguments{
\item{data}{a `tibble`}

\item{variables}{vector of varibles to caluclate distance}

\item{method}{method    one of the methods implemented in the stringdist package — "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex". See stringdist-metrics}

\item{clean_names}{if `TRUE` clean names}

\item{return_wide}{if `TRUE` widen long data}

\item{...}{other parameters passed to stringdist}
}
\value{
A tibble with string distance calculations
}
\description{
Calculate tidy string distance for a set of variables
}
\examples{
\dontrun{
tbl_proust <- tibble::tibble(person = c("Albertine", "Gilberte", "Odette"))
tbl_variable_stringdist(data = tbl_proust, variables = "person")
}
}
