% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyjoins.R
\name{tbl_variable_stringdist}
\alias{tbl_variable_stringdist}
\title{Variable String Distance}
\usage{
tbl_variable_stringdist(
  data,
  variables = NULL,
  method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  clean_names = T,
  return_wide = T,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Variable String Distance
}
