% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forebears.R
\name{us_political_affiliation}
\alias{us_political_affiliation}
\title{US political affiliations by last name}
\usage{
us_political_affiliation()
}
\value{

}
\description{
US political affiliations by last name
}
\examples{
us_political_affiliation()
}
